Implemented new lifecycle policy whereby all services can be accessed once FooService.register() has been called. Previously some services were available then and some where not available until ForPlay.run() had been called.
Note that in HtmlPlatform using setUseGL is no longer feasible, because the
decision to use GL or not is made before the caller has a reference to an
HtmlPlatform instance. We now accept a Mode enum which indicates whether we
should force the use of GL, Canvas, or autodetect the appropriate choice.