Merged various sample projects into single Showcase sample.
Each sample previously duplicated the ton of boilerplate needed to set up the
Core/Java, HTML, Flash and Android subprojects, and metaproject, and all the
associated wiring. There's nothing particularly illustrative about all that
duplicated boilerplate, and it makes it *way* harder to add a new sample.

This new Showcase takes care of all of the boilerplate once, and allows a new
sample to be added by simply implementing a single core class and adding it to
a list in Showcase.java.