SBT build configuration.
Yes, this is yet another build system, but SBT allows me to have magical
Eclipse-like cross project dependency calculations so that I can make a change
in any of my projects and have it magically reflected in the next test run of
whatever I'm working on.

At some point I'll write an SBT plugin that slurps metadata from POMs and that
will reduce the duplication somewhat.