Add accessors for homogenous json arrays
This adds accessors for homogenous arrays of all of the json types to
Json.Array and Json.Object. If you know your json data is of a single
type, this makes it way nicer to access an array. It's able to implement
Iterable since the type is known, and it can do some basic type checking
on the returned value.

It also makes Json.Object.getKeys slightly more pleasant to use since it
can return the iterable TypedArray<String> instead of just Array.