Introduced a mechanism for communicating the typed key.
The key code corresponds (roughly) to the key pressed on the keyboard. But the
actual unicode character that results is the product of the pressed key and
whatever keymap is active in the underlying operating system (and things are
even more complicated with mobile devices and their crazy input methods).
Fortunately, the underlying native events provide us with the unicode
character, we just need to pass it along.

I introduced onKeyTyped() which is issued in addition to onKeyDown(), both
because that's how Java does it (and it would be a major PITA to try to accrete
Java's two separate events into a single event), and because I think it's
cleaner to only receive an event for typed keys when there is actually a
unicode character to report. HTML, Android and Flash just issue a single event
(key down) which may or may not have a valid unicode character attached.