Use the JavaSound SPI mechanism to play MP3s (no more .wavs).
This obviates the need for JavAJLayerSound (which never really worked anyway).
I also substantially cleaned up the JavaSound implementation, which was doing
wacky things.

To be truly awesome, said implementation would use a clip pool to allow the
same sound to overlap with itself multiple times instead of stopping a playing
clip and starting it again, but that fancification is for another day.