Make Google App Engine Eclipse and Maven plugins play nicely together.
I am frankly appalled at what a mess this is. It should not be necessary to
download and unpack the GAE SDK when (the Google App Engine) Eclipse (Plugin)
has its own GAE SDK configurations. It should not be necessary for 50 lines of
boilerplate crap just to get Eclipse to invoke a lifecycle action on the
maven-gae-plugin. Much of the blame falls on Eclipse. If they would have just
gone the IDEA route to supporting Maven projects (i.e. let Maven do all the
work), most of this would be unnecessary.

Eclipse is a pox on humanity: it purports to increase developer productivity,
but in reality, crap like this consume ten times the effort that Eclipse
"saves" by letting you tab complete method names or run inane refactorings.