Further GL shader rearchitecture.
We remove the assumption that a shader will use a triangles and elements array
and let the shader make its own decisions. AbstractGLShader becomes
IndexedTrisShader to indicate that it forms a basis for shaders that plan to
use triangles and elements arrays.

We also introduce methods that provide the shaders for quad/tex, quad/color,
tris/tex and tris/color. The quads/tex and tris/tex may be the same shader, or
may be different (as will be the case for HTML-WebGL which is laden with
special needs).

In the case of WebGL, where elements arrays are non-performant, we can use
TRIANGLE_STRIP mode for our normal quad drawing (with a one-quad buffer). We
can use TRIANGLES mode, as usual, for triangle batch drawing, and automatically
expand the buffers (incurring the elements validation performance hit, but
there's no way to avoid it for triangle batches).

I think iOS supports QUADS; if so, we can use QUADS mode for quad drawing,
eliminating the elements array for the vast majority of our rendering.