Added notion of interactivity, a mechanism for adding listeners to layers.
A few things to note:
- Layers have interactive() and setInteractive().
- A layer's interactivity is false by default and automatically made true when
a listener is added to the layer; additionally a layer's parents are made
interactive when a layer becomes interactive.
- Interactivity and visibility are used by hitTest() which determines which, if
any, layer is hit by a point; invisible and non-interactive layers are omitted
from the search.

This will be used in a forthcoming commit to implement mouse/touch event
dispatch on layers.