We need to flush pending drawing before creating a framebuffer.
I could have just inserted a call to ctx.flush() in SurfaceGL.createTexture,
but then any future use of createFramebuffer would be susceptible to this same
bug. I could have also just documented that createFramebuffer has to flush
before binding its new framebuffer, but then a new platform could easily
accidentally omit this required flushing and have a bug.

So I took the more future-proof approach of having createFramebuffer be a
concrete method that does the flushing and calls a new createFramebufferImpl,
which is implemented by the backends and just has to create a framebuffer.