Added support for iPad as Phone mode.
In this mode, the iPad has logical resolution 384x512 and uses @2x images. The
Retina iPad also has logical resolution 384x512 and uses @4x images if
available, otherwise falls back to @2x (and even 1x, heaven forfend).

This makes life easier for phone games that want to use their Retina artwork
(or most of their Retina artwork) on the iPad, with only small UI rejiggering.

One can also supply 4 as a scale factor to the Java backend to emulate the
Retina iPad running in iPad as Phone mode. The screen is HUUUGE!