Work around odd Android Canvas drawRoundRect bug.
If the rect supplied to drawRoundRect has non-zero x or y, it causes the round
rectangle to be incorrectly rendered. So we simply translate the canvas, draw
using a rect at (0, 0) and then translate it back.

Surely someone would have discovered this bug already? Round rects are all the
rage in UIs.