Scale bitmaps down when using higher scale factor images.
For example, we might have a view scale factor of 1.5 (to make a WXGA device
look more like an iPhone) and the decide to use the iPhone Retina images on
said device. So we'll scale the 2x images down to 1.5x at load time, which
reduces GPU memory use and uses a much nicer scaling algorithm.