Added per-layer touch listeners.
This tracks an active layer per touch-id, so if you start a touch on a
particular layer, events related to that touch continue to be dispatched to
that layer until that touch ends. It is possible to have multiple touches being
dispatched to the same layer, in which case successive calls to the
LayerListener method take place on a single frame.

The iOS backend will need to do something about the fact that iOS doesn't
provide touch id. I seem to recall we figured out some what to deal with that,
but I'll have to dig it up.