Use the correct default framebuffer.
On iOS the default frame buffer is not zero. It uses a non-standard OpenGL
rendering model where you render everything into a framebuffer, which it then
composites using Core Animation. Previously I was steamrolling over that whole
requirement and just blasting everything into the main frame buffer. I'm kind
of surprised that it worked at all.

Unfortunately, this doesn't seem to fix the render-to-texture weirdness.