Reworked to render multiline text manually.
For some reason, using the CTFrame to render the text caused a larger and
larger inter-line spacing to be used for each line. I don't know if this is a
MonoTouch bug, or if I'm misusing the API, or what, but it's simpler in some
ways to just render each line individually at the proper position.