Further improvements to custom shaders.
We now properly inherit the shaders of our parent layer if we have no shader
configured ourselves.

It is also possible to configure the shader on a Surface, such that subsequent
drawing commands are done using that shader. ImmediateLayer pre-configures its
surface with the inherited shaders, SurfaceLayer does not. SurfaceLayer applies
the current shaders to its off-screen texture when rendering that, but when
creating the off-screen texture, there are no shaders to inherit. You can of
course configure shaders for use when rendering off-screen textures via
Surface.setShaders.