Added support for canceling pointer/touch interactions.
This happens naturally on iOS when, for example, an alert dialog is popped up
in the middle of a touch interaction. We now communicate that to the game. It
is also now possible to cancel pointer/touch interactions programmatically.
This is useful in situations where you have two overlapping interactions, like
button clicking and flick scrolling. A user might start a touch interaction,
which arms a button, but then flicks their finger up or down to scroll. In that
case, you want to cancel the touch interaction with the button and process the
flick to scroll interaction.

This is now possible via Touch#cancelLayerTouches and Pointer#cancelLayerDrags.
Note that we only cancel interactions happening on layers. The expectation is
that if you are processing interactions in your global Pointer/Touch listener,
you can handle programmatic cancelation yourself. If we canceled the global
interaction, there would be no way to complete the desired interaction (the
flick scroll for example).