Eliminated AbstractCachingAssets, and watching in AbstractAssets.
The base Assets implementations now just load assets, as advertised. If you
want to keep track of unloaded assets in the aggregate, use the newly added
WatchedAssets wrapper. If you want to cache your assets for the lifetime of
your game with no cache expiration policy, use CachingAssets. If you want to do
both, use both. Such is the magic of a wrapping delegate.