Great big async Assets revamp. Take note, Android/iOS users.
We now have:
- getImage       (always async)
- getImageSync   (always sync)
- getRemoteImage (always async)
- getSound       (always async)
- getText        (always async)
- getTextSync    (always sync)

There was no benefit to providing synchronous sound loading (you can do
everything with an async sound that you can with a sync sound), so I didn't
clutter the API with it.

Now if you want synchronous image/text loading, you have to explicitly use it,
and you can see in your code the places where you'll need to fix things if you
ever want to support an async-only platform like HTML5 or Flash.

Note that getImageSync and getTextSync throw UnsupportedOperation on
HTML5/Flash as they cannot be supported.