Standardize magnitude of mouse wheel event.
I think it would be OK to divide this by 120 instead. It certainly appears that in the lwjgl source,
they're scaling both Mac and Linux events by 120, and I *think* they're doing it to match what
windows does naturally - but that's not entirely clear, as that detail is hidden in the native
layer, and I'm no OpenGL wiz. This is safer and probably good enough for everybody.