Move our task to the back of the stack onBackPressed.
The default implementation finish()es our task which wreaks all sorts of havoc
for a PlayN game which really wants to be an app not a task. Moving to the back
yields (mostly) the same behavior but without shutting everything down.

Also protected-ized usePortraitOrientation as no one calls it outside
GameActivity.