New approach to handling rotation and views.
Now the IOSGameView is the root view controller's view, and we prevent the root
view controller from rotating at all (on iOS 5 or 6, that the mechanism for
controlling view rotation changed with iOS 6 was greatly confusing things).

So we have our root view controller (which things need) without all the
rotation fuckaroundery that usually comes with it. Yay. I thus also reinstated
the uiOverlay view because we no longer have any other view than the game view,
which is always unrotated.