What is this I don't even. Somehow this inversion is not needed.
I had everything working fine, tested on iOS 5.1 and 6.0 simulators. Then when
I tried playing my game, touch translation was hosed when the device was turned
upside down. I went back to test the PlayN test app and all of a sudden nothing
is working.

Furthermore, based on what I thought was a hard won understanding of what the
fuck was going on, an inverse transform *should* be needed here. I'm preventing
the root view controller (and hence the root view) from rotating when iOS asks
whether it should. So it should always think that it's in Portrait regardless
of the orientation of the device. Thus we should always get Portrait oriented
touches.

Now that I think about it, I reinstated the old uiOverlay *after* I did all my
testing. I also now recall some crazy-ass shit where, because touches were
passing "through" the uiOverlay (even though it's not handling them), they are
transformed based on the uiOverlay's transform (which is configured to account
for the rotation of the device). So since the uiOverlay is causing the touches
to be (inverse) transformed, we don't need to (inverse) transform them here.

Aaaanyway.