Allow non-trivial length audio to be played on the Java audio backend.
The javax.sampled.Clip has a very limited buffer size to the point that 30 second music loops would fail to load.

This adapts a BigClip class originally written by Andrew Thompson and posted to StackOverflow - it basically operates as a Clip but allowing larger buffer sizes. Some additional bugfixes were required to the version posted, and those are included here.