More asset related revampery.
Android handles the inclusion assets into the APK specially. Specifically, it
does not compress a variety of (already compressed) files when including them
in the APK. This is necessary so that it can later mmap() the data for those
files for use by, for example, MediaPlayer.

Because our assets are in the core project and then included into the android
project (and thence the APK) as a plain old dependency, aapt doesn't know they
should be treated as assets. So we pull our assets out into a separate
sub-project (like we do on iOS) and use a symlink in the standard Android
assets location (${projdir}/assets) to cause them to be picked up and treated
like proper assets.

The symlink will of course cause problems for Windows users, because Windows
(or Git on Windows, or something) is still stuck in the mid-20th century, but
we'll just have to find a way across that bridge, because having our assets
handled properly (instead of having to copy them to a temporary file system
before reading) is too juicy to pass up.