Update the archetype with separate assets submodule.
It turns out there's a Maven mojo that creates symlinks, so we might be able to
use the symlink approach without incredible pain. I'm sure Eclipse will choke
and gargle on the plugin and things won't work out of the box there, but when
have things ever worked out of the box on Eclipse?