Added frameInterval configuration.
This also switches the renderer to use CADisplayLink which syncs the updates to
the refresh rate of the display. Previously a timer was used which was kind of
wonky. Now you say you want 60FPS (the default, or 30FPS or 15FPS) and the OS
takes care of scheduling frame ticks.