Fix to make sure we really fill the expected pixel.
Stroking from 0, 0 to 1, 1 with the default stroke width (1?) will end up filling quite a bit
more than expected and ends up drawing a very fuzzy point.

I had originally been using setFillColor() and fillRect(0, 0, 1, 1) in Yohoho, and that worked
as expected. However, drawPoint() appears to be expected to use the stroke color, so we have to
do this context fiddling instead.