Allow canvas drawing to forgo interpolation.
Interpolation in canvas drawing was the source of a well hidden bug in yohoho; we're definitely
turning it off. It also helps to remove some of the "fuzzy" factor from yohoho art being shown
on a retina device, and interpolation of images when drawing to a canvas is slow.

I could see an argument being made to not even make this an option and just turn interpolation off
for everybody. The default interpolation probably isn't going to look how you want anyway. However,
this is somewhat similar to GLContext.setTextureFilter(), so I decided to make it an option.

I could also see an argument for making it a parameter (or maybe config as default with parameter
to override) on Graphics.createCanvasImage(), but that was a much bigger change I didn't feel like
making right now and I'm pretty sure we're going to want this off everywhere.