Workaround a bug in Intel graphics.
In IndexedTrisShaders, if we do the various attribute binding up in
activate(), Intel cards give broken behavior. To reproduce it, add three
layers:
 1) An ImageLayer from a CanvasImage (alternately use a SurfaceLayer)
 2) An ImageLayer from an assets-sourced Image
 3) Another ImageLayer from the same assets-sourced Image (subimages count)

Layers 2 & 3 wind up with Layer 1's geometry. If you only do layers 1 & 2,
everything's fine. If it's not a CanvasImage nor SurfaceLayer in position 1,
everything's fine. Very strange.

But we can just do the bind at the last possible chance instead of the first
possible, and everything's fine.