Most of SWT-hosted LWJGL backend.
I still need to wire up SWTKeyboard, and handle mouse wheel events.

If you want to try it out, change tests-java to depend on playn-swt-java and to
call SWTPlatform.register instead of JavaPlatform.register.

Also, I need to figure out if there's a non-crap way to include SWT native
libraries for all platforms. The person who packaged SWT for use by Maven seems
to have screwed the pooch in that regard. Right now it's only got a Mac 64-bit
depend, so you'll need to tweak that to test on other platforms.