Have Maven choose appropriate SWT for platform.
As the comment says, this is only to allow PlayN developers on all platforms to
have the right SWT jar chosen when they're running tests. If you want to deploy
an app using SWT (which is presumably the only reason you'd use this backend)
you need to include all the jars and have your app launcher put the appropriate
one on the classpath. Or you need per-platform installations (like Eclipse).
PITA PITA PITA.