Fixes to the shader clearing fix in 537b0936.
If we deactivate a shader, we need to null out curShader, otherwise that shader
might be immediately reactivated and think that its bits are already bound
(because it's still curShader).

This also forces unbinding when switching framebuffer, which seems to be
necessary to avoid funny business when rendering to a surface/texture.