Consolidated extras into core. Nixed color core vestige.
I decided to leave GLShader.Core around rather than merge it into GLShader
because the lifecycle of the GLShader is longer than the core. A single
GLShader may create and destroy its core multiple times over its lifecycle, and
it's nicer to have the core encapsulate all the bits that come and go rather
than have a bunch of mutable state in GLShader which follows implicit and
non-obvious rules about setting up and tearing down.