Fixed Android network mishandling null payload entities
When determining the payload strings from a successful HTTP request,
the previous implementation of AndroidNet would throw an
IllegalArgumentException and fail the request if the response entity is
null, such as in the case of a 204 status code.  Updated implementation
handles null response entities by invoking success on the callback with
the appropriate status code and null for the payload methods.