Nix IOSApplicationDelegate, use notifications.
This allows a PlayN app to be more easily embedded into larger iOS app. A
normal iOS app is going to want to have its own UIApplicationDelegate, and
having PlayN force the app to either use its delegate or manually add lifecycle
calls is non-awesome.

Fortunately, there's another way to get lifecycle callbacks which doesn't rely
on the UIApplicationDelegate. It does require the use of some generically typed
C# APIs which is ugly, but we'll survive.