Some tweaks to new RoboVM lifecycle &c.
Everything still now revolves around the (soon to be renamed)
RoboRootViewController, but we keep a RoboPlatform.register() method just for
symmetry with the other platforms and because "simple" clients don't need to
know about RoboRootViewController.

This exposes a bit more in the archetype stub (creating a window and showing
it), but that's something you have to do on iOS always, so it's not going to
change, and it's useful to understand what the basic bootstrap stuff looks like
on iOS so that you have a starting point if you want to dig deeper.