Fix more bugs, add RenderTarget.flip().
When we render to the default framebuffer, we want to flip the y-axis because
OpenGL thinks it's a math textbook and puts 0 at the bottom of the screen. But
every sane computer scientist knows that y is at the top of the screen, at
"framebuffer pointer plus zero".

When we render to a texture, OpenGL agrees with the computer scientists and 0 y
means "closest to the start of the texture memory", so we DON'T want to flip in
that case.

We used to hack around this by flipping the textures back again before
rendering them, but that's not as viable in the new world order where we invite
games to create textures however they like and we should do the right thing
with them.

So now we can tell a shader "render my shit upside down" or not. And when we
use a shader to render to a texture, we don't flip, and when we use it to
render to the default framebuffer, we do flip. No double flipping, no need to
flip out.