Moved createGradient() into Canvas.
You only ever use a Gradient with a Canvas, and unlike TextLayout, you're not
going to base the creation of your canvas on information returned from the
created Gradient.

It's possible that one will want to create a Gradient and keep it around, but
you can create a 1x1 canvas and use it to create your gradient(s). That's
basically what the HTML backend already has to do because the platform-specific
gradient object there has to be created via a Context2d which only a Canvas
has.

Now I want to get rid of Graphics.createFont() because that would reduce the
Graphics methods to creating textures, creating canvases, and laying out text,
which is pretty svelte. However, if I ditched createFont(), I'd probably have
to maintain a cache of PlayN to platform fonts, though frankly that's not a
horrible idea anyway...