No more platform specific Font. Woo!
Font is now just a container for configuration information. Each platform
maintains a cached mapping from Font to the platform font info (except HTML
where there is no platform font info, you just format it into CSS).

Java and Android separate size from typeface and style, so we're really just
caching typeface+style and we derive the appropriate style on the fly in the
layoutText() call (because it's comparatively cheap). iOS doesn't appear to
have a cheap way to convert a CTFont at one size into one at another, so we
just cache based on all three dimensions.

Any sane game is not going to use a zillion different combinations of font,
style and size anyway, so we're not looking at a big cache here.