Bundle a Bitmap and a Texture back together into Image.
When you don't want to worry about who creates a Texture for a Bitmap, but you
want to be sure that everyone is sharing the same Texture, you can use an
Image.

This subverts the memory savings that came from disconnecting Bitmap and
Texture, but sometimes you don't care about memory. It'll also be useful on
Android where we have to recreate our textures on GL context loss, so having
the image from whence they came will be helpful.