Vert, revert, treevert! Bitmap -> Image.
After starting to roll Texture + Bitmap + Image through TriplePlay, I realized
that there's no downside to having every Bitmap retain a reference to a
"default" texture. The texture does not retain a reference back to its image
(not yet anyway), and adding an extra field (two actually, but still) to Image
is hardly a terrible expense.

This allows us to have easy texture sharing for games that don't want to be
meticulous about texture management, without compromising things for games that
*do* want meticulous texture management. We also avoid an explosion of ways to
get a texture, which is nice.