Added RootLayer and Layer.state to track lifecycle.
One often wants to tie the lifecycle of other resources to a layer's
participation in the scene graph. When a layer goes into the scene graph, we
wire stuff up, when it comes out, we unwire stuff. That is now easy.

This introduces RootLayer which serves as an always ADDED group layer. When
things are added to a RootLayer (with any number of intervening parents), they
become ADDED, vice versa on removal. It is assumed that someone is painting the
root layer, by default that's SceneGame.