Take XY everywhere instead of IPoint.
We're relying on the fact that XY in PlayN pretty much always means a point. If
it meant a vector, then we'd be doing the wrong thing (because transforming a
vector is different than transforming a point).

But there's almost no reason why you'd be transforming a vector based on layer
transforms, that would be weird. So we're going to increase convenience by
allowing you to transform an Event or Interaction directly because they are
XYs.