Added support for LWJGL HiDPI mode.
This also changes some of the core stuff to allow the scale factor to be
changed on the fly. That happens when you draw a window from a Retina display
onto a non-Retina display, and vice versa.

We should probably emit a signal when the scale factor changes because a game
could conceivably want to regenerate graphics or reload assets for the right
display density, but that seems a bit crazy. For now it will go on using the
assets and bitmaps from the old scale factor, which will just be overkill if
you go from HiDPI to normal DPI, and which will look pixelly if you go the
other direction.

In any case, this is better than completely hosing the display.

Note: this also disables the manual scale factor twiddling that was previously
supported for the Java backend. I'll probably re-enable that, but I didn't want
it in there confusing things when I was trying to get proper scale factor stuff
working.