Java backend cleanup.
This is in preparation for porting Ruslan's JOGL-based backend to 2.0 and
offering it as a third (!) Java backend.

Now we have:
- java-base  - stuff shared by all Java backends
- java-lwjgl - a bare LWJGL backend
- java-swt   - an LWJGL+SWT backend

Shortly we'll also have java-jogl which will be a JOGL+AWT backend.

The archetype assumes you want to use the bare LWJGL backend. You can tweak
your (one) game/java module to use LWJGL+SWT or JOGL if you prefer that one, or
copy it if you really want to support multiple Java backends for your game.

Projects already using PlayN 2.0 will need to change their java/pom.xml to
reference playn-java-lwjgl instead of playn-java and their Java bootstrap to
create an LWJGLPlatform instead of a JavaPlatform.