Added above & below capture modes.
Previously, you could capture an interaction and subsequent events would go
only to the capturing layer. But sometimes you want events to go from the hit
layer *up* to the capturing layer, or from the capturing layer up to the root.
So now we support that.

This also fixes some niggles with nested event dispatch (which only happens
when dispatching the cancel triggered by capture).