Revert Graphics.setSize().
There is no cross-platform way to "set the size" of the view. Some platforms
support setting the size, some do not. Those that do not can expose the
mechanism for telling PlayN when the size changed in the event that special
users want to make use of PlayN without the standard view size, but we should
not expose a public, cross platform method that "looks" like it will change the
size of the main game view, because that will just confuse people.