Step down in integers from the native scale factor.
Many Android devices have 3x pixel density, and it's weird to go 3x, 1.5x, 1x.
So instead we go 3x, 2x, 1x. This means that a Retina iPad in
pretend-to-be-iPhone mode would go 4x, 3x, 2x, 1x, instead of 4x, 2x, 1x, but
that's not a huge problem. It's just one more failed fstat when loading an
asset, assuming it's using 2x assets. Plus that's often happening on a
background thread.