We always want the Canvas to be "full size".
So if you're running on a HiDPI display with 2 physical pixels for every one
logical pixel, we always want the canvas to be 2x the "logical" width and
height specified for your game.

All the rest of the PlayN code will then do the right thing with the "logical"
scale factor that you configure, in this case in HtmlPlatform.Config.

If you leave the logical scale factor at the default, it will match the device
scale factor and you'll be working in logical pixels and everything will just
be sharper.

If you set the logical scale factor to 1, then you are now working in a one to
one physical to logical pixel world and it's up to you to decide how to scale
things to use all those pixels.