Remove APK expansion file support.
This required the apk-expansion library which Google does not publish to
Maven Central. We had published it to our own hacky Maven repo ages ago,
but that went away when Google Code went away.

Since this was the last last vestige to need artifacts from that repo, I
just opted to remove support for APK expansion files from the baseline
repo and make some minor restructuring to the AndroidAssets to make it
easy for a game to use APK expansion files by plugging in a custom
AndroidAssets.AssetSource.

I'll add a cookbook entry on how to load assets from an APK expansion
file. If a game is built via Gradle (which supports Android's weird
non-standard dependency mechanisms better), it will still be relatively
easy to make this work.

I also plan (in my infinite spare time) to convert one of the sample
projects over to a Gradle build so that if people want to build their
games via Gradle, they have an example to follow in setting up their
build. Another thing to do in my infinite spare time.