Added Net.Response.payloadImage().
Creates an image from the results of an HTTP request. This API is not supported
on the HTML backend. It could almost work but we'd need to add an extra API to
say "give me the binary response data" because HTML is awesome and forces you
to decide if you want a string or binary data before you issue your request.